# apps/webapp/src/assets/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/v1/auth/register":{"post":{"operationId":"AuthController_register_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Successfully registered and logged in."}},"summary":"Register a new user","tags":["Auth"]}},"/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Successfully logged in."}},"summary":"Login with email and password","tags":["Auth"]}},"/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshDto"}}}},"responses":{"200":{"description":"Successfully refreshed tokens."}},"summary":"Refresh access and refresh tokens","tags":["Auth"]}},"/v1/auth/logout":{"post":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"200":{"description":"Successfully logged out."}},"security":[{"bearer":[]}],"summary":"Logout and revoke session","tags":["Auth"]}},"/v1/auth/me":{"get":{"operationId":"AuthController_getMe_v1","parameters":[],"responses":{"200":{"description":"Returns the current user payload."}},"security":[{"bearer":[]}],"summary":"Get current logged in user details","tags":["Auth"]}},"/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"summary":"Request password reset link","tags":["Auth"]}},"/v1/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"summary":"Reset password using token","tags":["Auth"]}},"/v1/auth/change-password":{"post":{"operationId":"AuthController_changePassword_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Change password while logged in","tags":["Auth"]}},"/v1/auth/verify-email":{"post":{"operationId":"AuthController_verifyEmail_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDto"}}}},"responses":{"200":{"description":""}},"summary":"Verify user email","tags":["Auth"]}}},"info":{"title":"SECURE.ID API","description":"# ERP-IAM SecureID API Documentation\n\nWelcome to the **SecureID** API. This service handles all authentication and authorization for the ERP-IAM ecosystem.\n\n## Key Features\n\n- **Fastify-powered**: High performance, low overhead.\n- **RS256 JWT**: Secure token-based authentication.\n- **RBAC**: Robust Role-Based Access Control.\n- **Audit Logging**: All sensitive actions are logged to PostgreSQL.\n\n## Getting Started\n\nTo use the protected endpoints, you must first register or login to receive a Bearer token. Use the `Authorize` button below to include the token in your requests.\n\n---\n\n_Powered by NestJS & Scalar Reference_\n","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"newuser@erpiam.local"},"password":{"type":"string","example":"secret123"},"name":{"type":"string","example":"John Doe"},"companyName":{"type":"string","example":"Acme Corp"}},"required":["email","password","name","companyName"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@erpiam.local"},"password":{"type":"string","example":"secret123"}},"required":["email","password"]},"RefreshDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"token":{"type":"string"},"newPassword":{"type":"string"}},"required":["token","newPassword"]},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"VerifyEmailDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]}}}}